Split up the db_accessor header
ClangBuildAnalyzer
  Parsing (frontend):         1616.3 s
  Codegen & opts (backend):    424.9 s
