+++ src/query/context.hpp
@@ -59,24 +59,9 @@ struct EvaluationContext {
   Scope scope{};
 };
 
-inline std::vector<storage::PropertyId> NamesToProperties(const std::vector<std::string> &property_names,
-                                                          DbAccessor *dba) {
-  std::vector<storage::PropertyId> properties;
-  properties.reserve(property_names.size());
-  for (const auto &name : property_names) {
-    properties.push_back(dba->NameToProperty(name));
-  }
-  return properties;
-}
+std::vector<storage::PropertyId> NamesToProperties(const std::vector<std::string> &property_names, DbAccessor *dba);
 
-inline std::vector<storage::LabelId> NamesToLabels(const std::vector<std::string> &label_names, DbAccessor *dba) {
-  std::vector<storage::LabelId> labels;
-  labels.reserve(label_names.size());
-  for (const auto &name : label_names) {
-    labels.push_back(dba->NameToLabel(name));
-  }
-  return labels;
-}
+std::vector<storage::LabelId> NamesToLabels(const std::vector<std::string> &label_names, DbAccessor *dba);
 
 struct ExecutionContext {
   DbAccessor *db_accessor{nullptr};
