+++ src/query/db_accessor.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Memgraph Ltd.
+// Copyright 2024 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -163,7 +163,7 @@ storage::Result<EdgeVertexAccessorResult> SubgraphVertexAccessor::OutEdges(stora
   std::vector<EdgeAccessor> resulting_edges;
   resulting_edges.reserve(filteredOutEdges.size());
   std::ranges::transform(filteredOutEdges, std::back_inserter(resulting_edges),
-                         [](auto const &edge) { return VertexAccessor::MakeEdgeAccessor(edge); });
+                         [](auto const &edge) { return EdgeAccessor(edge); });
 
   return EdgeVertexAccessorResult{.edges = std::move(resulting_edges), .expanded_count = maybe_edges->expanded_count};
 }
@@ -185,7 +185,7 @@ storage::Result<EdgeVertexAccessorResult> SubgraphVertexAccessor::InEdges(storag
   std::vector<EdgeAccessor> resulting_edges;
   resulting_edges.reserve(filteredOutEdges.size());
   std::ranges::transform(filteredOutEdges, std::back_inserter(resulting_edges),
-                         [](auto const &edge) { return VertexAccessor::MakeEdgeAccessor(edge); });
+                         [](auto const &edge) { return EdgeAccessor(edge); });
 
   return EdgeVertexAccessorResult{.edges = std::move(resulting_edges), .expanded_count = maybe_edges->expanded_count};
 }
