+++ src/query/db_accessor.hpp
@@ -18,8 +18,10 @@
 #include <cppitertools/imap.hpp>
 
 #include "memory/query_memory_control.hpp"
+#include "query/edge_accessor.hpp"
 #include "query/exceptions.hpp"
 #include "query/hops_limit.hpp"
+#include "query/vertex_accessor.hpp"
 #include "storage/v2/constraints/type_constraints.hpp"
 #include "storage/v2/edge_accessor.hpp"
 #include "storage/v2/id_types.hpp"
@@ -35,224 +37,9 @@
 #include "utils/result.hpp"
 #include "utils/variant_helpers.hpp"
 
-// TODO: split into smaller headers
-
 namespace memgraph::query {
 
 class Graph;
-class VertexAccessor;
-
-class EdgeAccessor final {
- public:
-  storage::EdgeAccessor impl_;
-
-  explicit EdgeAccessor(storage::EdgeAccessor impl) : impl_(std::move(impl)) {}
-
-  bool IsDeleted() const { return impl_.IsDeleted(); }
-
-  bool IsVisible(storage::View view) const { return impl_.IsVisible(view); }
-
-  storage::EdgeTypeId EdgeType() const { return impl_.EdgeType(); }
-
-  auto Properties(storage::View view) const { return impl_.Properties(view); }
-
-  storage::Result<storage::PropertyValue> GetProperty(storage::View view, storage::PropertyId key) const {
-    return impl_.GetProperty(key, view);
-  }
-
-  storage::Result<uint64_t> GetPropertySize(storage::PropertyId key, storage::View view) const {
-    return impl_.GetPropertySize(key, view);
-  }
-
-  storage::Result<storage::PropertyValue> SetProperty(storage::PropertyId key, const storage::PropertyValue &value) {
-    return impl_.SetProperty(key, value);
-  }
-
-  storage::Result<bool> InitProperties(const std::map<storage::PropertyId, storage::PropertyValue> &properties) {
-    return impl_.InitProperties(properties);
-  }
-
-  storage::Result<std::vector<std::tuple<storage::PropertyId, storage::PropertyValue, storage::PropertyValue>>>
-  UpdateProperties(std::map<storage::PropertyId, storage::PropertyValue> &properties) const {
-    return impl_.UpdateProperties(properties);
-  }
-
-  storage::Result<storage::PropertyValue> RemoveProperty(storage::PropertyId key) {
-    return SetProperty(key, storage::PropertyValue());
-  }
-
-  storage::Result<std::map<storage::PropertyId, storage::PropertyValue>> ClearProperties() {
-    return impl_.ClearProperties();
-  }
-
-  VertexAccessor To() const;
-
-  VertexAccessor From() const;
-
-  /// When edge is deleted and you are accessing To vertex
-  /// for_deleted_ flag will in this case be updated properly
-  VertexAccessor DeletedEdgeToVertex() const;
-
-  /// When edge is deleted and you are accessing From vertex
-  /// for_deleted_ flag will in this case be updated properly
-  VertexAccessor DeletedEdgeFromVertex() const;
-
-  bool IsCycle() const;
-
-  int64_t CypherId() const { return impl_.Gid().AsInt(); }
-
-  storage::Gid Gid() const noexcept { return impl_.Gid(); }
-
-  bool operator==(const EdgeAccessor &e) const noexcept { return impl_ == e.impl_; }
-
-  bool operator!=(const EdgeAccessor &e) const noexcept { return !(*this == e); }
-};
-
-struct EdgeVertexAccessorResult {
-  std::vector<EdgeAccessor> edges;
-  int64_t expanded_count;
-};
-
-class VertexAccessor final {
- public:
-  storage::VertexAccessor impl_;
-
-  static EdgeAccessor MakeEdgeAccessor(const storage::EdgeAccessor impl) { return EdgeAccessor(impl); }
-
-  explicit VertexAccessor(storage::VertexAccessor impl) : impl_(impl) {}
-
-  bool IsVisible(storage::View view) const { return impl_.IsVisible(view); }
-
-  auto Labels(storage::View view) const { return impl_.Labels(view); }
-
-  storage::Result<bool> AddLabel(storage::LabelId label) { return impl_.AddLabel(label); }
-
-  storage::Result<bool> RemoveLabel(storage::LabelId label) { return impl_.RemoveLabel(label); }
-
-  storage::Result<bool> HasLabel(storage::View view, storage::LabelId label) const {
-    return impl_.HasLabel(label, view);
-  }
-
-  auto Properties(storage::View view) const { return impl_.Properties(view); }
-
-  storage::Result<storage::PropertyValue> GetProperty(storage::View view, storage::PropertyId key) const {
-    return impl_.GetProperty(key, view);
-  }
-
-  storage::Result<uint64_t> GetPropertySize(storage::PropertyId key, storage::View view) const {
-    return impl_.GetPropertySize(key, view);
-  }
-
-  storage::Result<storage::PropertyValue> SetProperty(storage::PropertyId key, const storage::PropertyValue &value) {
-    return impl_.SetProperty(key, value);
-  }
-
-  storage::Result<bool> InitProperties(const std::map<storage::PropertyId, storage::PropertyValue> &properties) {
-    return impl_.InitProperties(properties);
-  }
-
-  storage::Result<std::vector<std::tuple<storage::PropertyId, storage::PropertyValue, storage::PropertyValue>>>
-  UpdateProperties(std::map<storage::PropertyId, storage::PropertyValue> &properties) const {
-    return impl_.UpdateProperties(properties);
-  }
-
-  storage::Result<storage::PropertyValue> RemoveProperty(storage::PropertyId key) {
-    return SetProperty(key, storage::PropertyValue());
-  }
-
-  storage::Result<std::map<storage::PropertyId, storage::PropertyValue>> ClearProperties() {
-    return impl_.ClearProperties();
-  }
-
-  storage::Result<EdgeVertexAccessorResult> InEdges(storage::View view,
-                                                    const std::vector<storage::EdgeTypeId> &edge_types,
-                                                    query::HopsLimit *hops_limit = nullptr) const {
-    auto maybe_result = impl_.InEdges(view, edge_types, nullptr, hops_limit);
-    if (maybe_result.HasError()) return maybe_result.GetError();
-
-    std::vector<EdgeAccessor> edges;
-    edges.reserve((*maybe_result).edges.size());
-    std::ranges::transform((*maybe_result).edges, std::back_inserter(edges),
-                           [](auto const &edge) { return EdgeAccessor(edge); });
-
-    return EdgeVertexAccessorResult{.edges = edges, .expanded_count = (*maybe_result).expanded_count};
-  }
-
-  storage::Result<EdgeVertexAccessorResult> InEdges(storage::View view) const { return InEdges(view, {}); }
-
-  storage::Result<EdgeVertexAccessorResult> InEdges(storage::View view,
-                                                    const std::vector<storage::EdgeTypeId> &edge_types,
-                                                    const VertexAccessor &dest,
-                                                    query::HopsLimit *hops_limit = nullptr) const {
-    auto maybe_result = impl_.InEdges(view, edge_types, &dest.impl_, hops_limit);
-    if (maybe_result.HasError()) return maybe_result.GetError();
-
-    std::vector<EdgeAccessor> edges;
-    edges.reserve((*maybe_result).edges.size());
-    std::ranges::transform((*maybe_result).edges, std::back_inserter(edges),
-                           [](auto const &edge) { return EdgeAccessor(edge); });
-
-    return EdgeVertexAccessorResult{.edges = edges, .expanded_count = (*maybe_result).expanded_count};
-  }
-
-  storage::Result<EdgeVertexAccessorResult> OutEdges(storage::View view,
-                                                     const std::vector<storage::EdgeTypeId> &edge_types,
-                                                     query::HopsLimit *hops_limit = nullptr) const {
-    auto maybe_result = impl_.OutEdges(view, edge_types, nullptr, hops_limit);
-    if (maybe_result.HasError()) return maybe_result.GetError();
-
-    std::vector<EdgeAccessor> edges;
-    edges.reserve((*maybe_result).edges.size());
-    std::ranges::transform((*maybe_result).edges, std::back_inserter(edges),
-                           [](auto const &edge) { return EdgeAccessor(edge); });
-
-    return EdgeVertexAccessorResult{.edges = edges, .expanded_count = (*maybe_result).expanded_count};
-  }
-
-  storage::Result<EdgeVertexAccessorResult> OutEdges(storage::View view) const { return OutEdges(view, {}); }
-
-  storage::Result<EdgeVertexAccessorResult> OutEdges(storage::View view,
-                                                     const std::vector<storage::EdgeTypeId> &edge_types,
-                                                     const VertexAccessor &dest,
-                                                     query::HopsLimit *hops_limit = nullptr) const {
-    auto maybe_result = impl_.OutEdges(view, edge_types, &dest.impl_, hops_limit);
-    if (maybe_result.HasError()) return maybe_result.GetError();
-
-    std::vector<EdgeAccessor> edges;
-    edges.reserve((*maybe_result).edges.size());
-    std::ranges::transform((*maybe_result).edges, std::back_inserter(edges),
-                           [](auto const &edge) { return EdgeAccessor(edge); });
-
-    return EdgeVertexAccessorResult{.edges = edges, .expanded_count = (*maybe_result).expanded_count};
-  }
-
-  storage::Result<size_t> InDegree(storage::View view) const { return impl_.InDegree(view); }
-
-  storage::Result<size_t> OutDegree(storage::View view) const { return impl_.OutDegree(view); }
-
-  int64_t CypherId() const { return impl_.Gid().AsInt(); }
-
-  storage::Gid Gid() const noexcept { return impl_.Gid(); }
-
-  bool operator==(const VertexAccessor &v) const noexcept {
-    static_assert(noexcept(impl_ == v.impl_));
-    return impl_ == v.impl_;
-  }
-
-  bool operator!=(const VertexAccessor &v) const noexcept { return !(*this == v); }
-};
-
-inline VertexAccessor EdgeAccessor::To() const { return VertexAccessor(impl_.ToVertex()); }
-
-inline VertexAccessor EdgeAccessor::From() const { return VertexAccessor(impl_.FromVertex()); }
-
-inline VertexAccessor EdgeAccessor::DeletedEdgeToVertex() const { return VertexAccessor(impl_.DeletedEdgeToVertex()); }
-
-inline VertexAccessor EdgeAccessor::DeletedEdgeFromVertex() const {
-  return VertexAccessor(impl_.DeletedEdgeFromVertex());
-}
-
-inline bool EdgeAccessor::IsCycle() const { return To() == From(); }
 
 class SubgraphVertexAccessor final {
  public:
@@ -308,19 +95,6 @@ class SubgraphVertexAccessor final {
   VertexAccessor GetVertexAccessor() const;
 };
 }  // namespace memgraph::query
-
-namespace std {
-template <>
-struct hash<memgraph::query::VertexAccessor> {
-  size_t operator()(const memgraph::query::VertexAccessor &v) const { return std::hash<decltype(v.impl_)>{}(v.impl_); }
-};
-
-template <>
-struct hash<memgraph::query::EdgeAccessor> {
-  size_t operator()(const memgraph::query::EdgeAccessor &e) const { return std::hash<decltype(e.impl_)>{}(e.impl_); }
-};
-
-}  // namespace std
 
 namespace memgraph::query {
