+++ src/query/interpret/eval.hpp
@@ -481,396 +481,9 @@ class ExpressionEvaluator : public ExpressionVisitor<TypedValue> {
     return TypedValue(value.IsNull(), ctx_->memory);
   }
 
-  TypedValue Visit(PropertyLookup &property_lookup) override {
-    ReferenceExpressionEvaluator referenceExpressionEvaluator(frame_, symbol_table_, ctx_);
+  TypedValue Visit(PropertyLookup &property_lookup) override;
 
-    TypedValue *expression_result_ptr = property_lookup.expression_->Accept(referenceExpressionEvaluator);
-    TypedValue expression_result;
-
-    if (nullptr == expression_result_ptr) {
-      expression_result = property_lookup.expression_->Accept(*this);
-      expression_result_ptr = &expression_result;
-    }
-    auto maybe_date = [this](const auto &date, const auto &prop_name) -> std::optional<TypedValue> {
-      if (prop_name == "year") {
-        return TypedValue(date.year, ctx_->memory);
-      }
-      if (prop_name == "month") {
-        return TypedValue(date.month, ctx_->memory);
-      }
-      if (prop_name == "day") {
-        return TypedValue(date.day, ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_local_time = [this](const auto &lt, const auto &prop_name) -> std::optional<TypedValue> {
-      if (prop_name == "hour") {
-        return TypedValue(lt.hour, ctx_->memory);
-      }
-      if (prop_name == "minute") {
-        return TypedValue(lt.minute, ctx_->memory);
-      }
-      if (prop_name == "second") {
-        return TypedValue(lt.second, ctx_->memory);
-      }
-      if (prop_name == "millisecond") {
-        return TypedValue(lt.millisecond, ctx_->memory);
-      }
-      if (prop_name == "microsecond") {
-        return TypedValue(lt.microsecond, ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_duration = [this](const auto &dur, const auto &prop_name) -> std::optional<TypedValue> {
-      if (prop_name == "day") {
-        return TypedValue(dur.Days(), ctx_->memory);
-      }
-      if (prop_name == "hour") {
-        return TypedValue(dur.SubDaysAsHours(), ctx_->memory);
-      }
-      if (prop_name == "minute") {
-        return TypedValue(dur.SubDaysAsMinutes(), ctx_->memory);
-      }
-      if (prop_name == "second") {
-        return TypedValue(dur.SubDaysAsSeconds(), ctx_->memory);
-      }
-      if (prop_name == "millisecond") {
-        return TypedValue(dur.SubDaysAsMilliseconds(), ctx_->memory);
-      }
-      if (prop_name == "microsecond") {
-        return TypedValue(dur.SubDaysAsMicroseconds(), ctx_->memory);
-      }
-      if (prop_name == "nanosecond") {
-        return TypedValue(dur.SubDaysAsNanoseconds(), ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_zoned_date_time = [this](const auto &zdt, const auto &prop_name) -> std::optional<TypedValue> {
-      if (prop_name == "year") {
-        return TypedValue(zdt.LocalYear(), ctx_->memory);
-      }
-      if (prop_name == "month") {
-        return TypedValue(utils::MemcpyCast<int64_t>(zdt.LocalMonth()), ctx_->memory);
-      }
-      if (prop_name == "day") {
-        return TypedValue(utils::MemcpyCast<int64_t>(zdt.LocalDay()), ctx_->memory);
-      }
-      if (prop_name == "hour") {
-        return TypedValue(zdt.LocalHour(), ctx_->memory);
-      }
-      if (prop_name == "minute") {
-        return TypedValue(zdt.LocalMinute(), ctx_->memory);
-      }
-      if (prop_name == "second") {
-        return TypedValue(zdt.LocalSecond(), ctx_->memory);
-      }
-      if (prop_name == "millisecond") {
-        return TypedValue(zdt.LocalMillisecond(), ctx_->memory);
-      }
-      if (prop_name == "microsecond") {
-        return TypedValue(zdt.LocalMicrosecond(), ctx_->memory);
-      }
-      if (prop_name == "timezone") {
-        return TypedValue(zdt.GetTimezone().ToString(), ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_point2d = [this](const auto &point_2d, const auto &prop_name) -> std::optional<TypedValue> {
-      auto is_wgs = point_2d.crs() == storage::CoordinateReferenceSystem::WGS84_2d;
-      if (prop_name == "x") {
-        return TypedValue(point_2d.x(), ctx_->memory);
-      }
-      if (prop_name == "longitude") {
-        if (!is_wgs) throw QueryRuntimeException("Use x instead of longitude for cartesian point types");
-        return TypedValue(point_2d.x(), ctx_->memory);
-      }
-      if (prop_name == "y") {
-        return TypedValue(point_2d.y(), ctx_->memory);
-      }
-      if (prop_name == "latitude") {
-        if (!is_wgs) throw QueryRuntimeException("Use y instead of latitude for cartesian point types");
-        return TypedValue(point_2d.y(), ctx_->memory);
-      }
-      if (prop_name == "crs") {
-        return TypedValue(storage::CrsToString(point_2d.crs()), ctx_->memory);
-      }
-      if (prop_name == "srid") {
-        return TypedValue(storage::CrsToSrid(point_2d.crs()).value_of(), ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_point3d = [this](const auto &point_3d, const auto &prop_name) -> std::optional<TypedValue> {
-      auto is_wgs = point_3d.crs() == storage::CoordinateReferenceSystem::WGS84_3d;
-      if (prop_name == "x") {
-        return TypedValue(point_3d.x(), ctx_->memory);
-      }
-      if (prop_name == "longitude") {
-        if (!is_wgs) throw QueryRuntimeException("Use x instead of longitude for cartesian point types");
-        return TypedValue(point_3d.x(), ctx_->memory);
-      }
-      if (prop_name == "y") {
-        return TypedValue(point_3d.y(), ctx_->memory);
-      }
-      if (prop_name == "latitude") {
-        if (!is_wgs) throw QueryRuntimeException("Use y instead of latitude for cartesian point types");
-        return TypedValue(point_3d.y(), ctx_->memory);
-      }
-      if (prop_name == "z") {
-        return TypedValue(point_3d.z(), ctx_->memory);
-      }
-      if (prop_name == "height") {
-        if (!is_wgs) throw QueryRuntimeException("Use z instead of height for cartesian point types");
-        return TypedValue(point_3d.z(), ctx_->memory);
-      }
-      if (prop_name == "crs") {
-        return TypedValue(storage::CrsToString(point_3d.crs()), ctx_->memory);
-      }
-      if (prop_name == "srid") {
-        return TypedValue(storage::CrsToSrid(point_3d.crs()).value_of(), ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    auto maybe_graph = [this](const auto &graph, const auto &prop_name) -> std::optional<TypedValue> {
-      if (prop_name == "nodes") {
-        utils::pmr::vector<TypedValue> vertices(ctx_->memory);
-        vertices.reserve(graph.vertices().size());
-        for (const auto &v : graph.vertices()) {
-          vertices.emplace_back(TypedValue(v, ctx_->memory));
-        }
-        return TypedValue(vertices, ctx_->memory);
-      }
-      if (prop_name == "edges") {
-        utils::pmr::vector<TypedValue> edges(ctx_->memory);
-        edges.reserve(graph.edges().size());
-        for (const auto &e : graph.edges()) {
-          edges.emplace_back(TypedValue(e, ctx_->memory));
-        }
-        return TypedValue(edges, ctx_->memory);
-      }
-      return std::nullopt;
-    };
-    switch (expression_result_ptr->type()) {
-      case TypedValue::Type::Null:
-        return TypedValue(ctx_->memory);
-      case TypedValue::Type::Vertex:
-        if (property_lookup.evaluation_mode_ == PropertyLookup::EvaluationMode::GET_ALL_PROPERTIES) {
-          auto symbol_pos = static_cast<Identifier *>(property_lookup.expression_)->symbol_pos_;
-          if (!property_lookup_cache_.contains(symbol_pos)) {
-            property_lookup_cache_.emplace(symbol_pos, GetAllProperties(expression_result_ptr->ValueVertex()));
-          }
-
-          auto property_id = ctx_->properties[property_lookup.property_.ix];
-          if (property_lookup_cache_[symbol_pos].contains(property_id)) {
-            return TypedValue(property_lookup_cache_[symbol_pos][property_id], ctx_->memory);
-          }
-          return TypedValue(ctx_->memory);
-        } else {
-          return TypedValue(GetProperty(expression_result_ptr->ValueVertex(), property_lookup.property_), ctx_->memory);
-        }
-      case TypedValue::Type::Edge:
-        if (property_lookup.evaluation_mode_ == PropertyLookup::EvaluationMode::GET_ALL_PROPERTIES) {
-          auto symbol_pos = static_cast<Identifier *>(property_lookup.expression_)->symbol_pos_;
-          if (!property_lookup_cache_.contains(symbol_pos)) {
-            property_lookup_cache_.emplace(symbol_pos, GetAllProperties(expression_result_ptr->ValueEdge()));
-          }
-
-          auto property_id = ctx_->properties[property_lookup.property_.ix];
-          if (property_lookup_cache_[symbol_pos].contains(property_id)) {
-            return TypedValue(property_lookup_cache_[symbol_pos][property_id], ctx_->memory);
-          }
-          return TypedValue(ctx_->memory);
-        } else {
-          return TypedValue(GetProperty(expression_result_ptr->ValueEdge(), property_lookup.property_), ctx_->memory);
-        }
-      case TypedValue::Type::Map: {
-        auto &map = expression_result_ptr->ValueMap();
-        auto found = map.find(property_lookup.property_.name.c_str());
-        if (found == map.end()) return TypedValue(ctx_->memory);
-        return TypedValue(found->second, ctx_->memory);
-      }
-      case TypedValue::Type::Duration: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &dur = expression_result_ptr->ValueDuration();
-        if (auto dur_field = maybe_duration(dur, prop_name); dur_field) {
-          return TypedValue(*dur_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for Duration", prop_name);
-      }
-      case TypedValue::Type::Date: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &date = expression_result_ptr->ValueDate();
-        if (auto date_field = maybe_date(date, prop_name); date_field) {
-          return TypedValue(*date_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for Date", prop_name);
-      }
-      case TypedValue::Type::LocalTime: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &lt = expression_result_ptr->ValueLocalTime();
-        if (auto lt_field = maybe_local_time(lt, prop_name); lt_field) {
-          return std::move(*lt_field);
-        }
-        throw QueryRuntimeException("Invalid property name {} for LocalTime", prop_name);
-      }
-      case TypedValue::Type::LocalDateTime: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &ldt = expression_result_ptr->ValueLocalDateTime();
-        if (auto date_field = maybe_date(ldt.date(), prop_name); date_field) {
-          return std::move(*date_field);
-        }
-        if (auto lt_field = maybe_local_time(ldt.local_time(), prop_name); lt_field) {
-          return TypedValue(*lt_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for LocalDateTime", prop_name);
-      }
-      case TypedValue::Type::ZonedDateTime: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &zdt = expression_result_ptr->ValueZonedDateTime();
-        if (auto zdt_field = maybe_zoned_date_time(zdt, prop_name); zdt_field) {
-          return TypedValue(*zdt_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for ZonedDateTime", prop_name);
-      }
-      case TypedValue::Type::Point2d: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &point_2d = expression_result_ptr->ValuePoint2d();
-        if (auto point_2d_field = maybe_point2d(point_2d, prop_name); point_2d_field) {
-          return TypedValue(*point_2d_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for Point2d", prop_name);
-      }
-      case TypedValue::Type::Point3d: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &point_3d = expression_result_ptr->ValuePoint3d();
-        if (auto point_3d_field = maybe_point3d(point_3d, prop_name); point_3d_field) {
-          return TypedValue(*point_3d_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for Point3d", prop_name);
-      }
-      case TypedValue::Type::Graph: {
-        const auto &prop_name = property_lookup.property_.name;
-        const auto &graph = expression_result_ptr->ValueGraph();
-        if (auto graph_field = maybe_graph(graph, prop_name); graph_field) {
-          return TypedValue(*graph_field, ctx_->memory);
-        }
-        throw QueryRuntimeException("Invalid property name {} for Graph", prop_name);
-      }
-      default:
-        throw QueryRuntimeException(
-            "Only nodes, edges, maps, temporal types and graphs have properties to be looked up.");
-    }
-  }
-
-  TypedValue Visit(AllPropertiesLookup &all_properties_lookup) override {
-    TypedValue::TMap result(ctx_->memory);
-
-    auto expression_result = all_properties_lookup.expression_->Accept(*this);
-    switch (expression_result.type()) {
-      case TypedValue::Type::Null:
-        return TypedValue(ctx_->memory);
-      case TypedValue::Type::Vertex: {
-        for (const auto properties = *expression_result.ValueVertex().Properties(view_);
-             const auto &[property_id, value] : properties) {
-          result.emplace(dba_->PropertyToName(property_id), value);
-        }
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Edge: {
-        for (const auto properties = *expression_result.ValueEdge().Properties(view_);
-             const auto &[property_id, value] : properties) {
-          result.emplace(dba_->PropertyToName(property_id), value);
-        }
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Map: {
-        for (auto &[name, value] : expression_result.ValueMap()) {
-          result.emplace(name, value);
-        }
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Duration: {
-        const auto &dur = expression_result.ValueDuration();
-        result.emplace("day", TypedValue(dur.Days(), ctx_->memory));
-        result.emplace("hour", TypedValue(dur.SubDaysAsHours(), ctx_->memory));
-        result.emplace("minute", TypedValue(dur.SubDaysAsMinutes(), ctx_->memory));
-        result.emplace("second", TypedValue(dur.SubDaysAsSeconds(), ctx_->memory));
-        result.emplace("millisecond", TypedValue(dur.SubDaysAsMilliseconds(), ctx_->memory));
-        result.emplace("microseconds", TypedValue(dur.SubDaysAsMicroseconds(), ctx_->memory));
-        result.emplace("nanoseconds", TypedValue(dur.SubDaysAsNanoseconds(), ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Date: {
-        const auto &date = expression_result.ValueDate();
-        result.emplace("year", TypedValue(date.year, ctx_->memory));
-        result.emplace("month", TypedValue(date.month, ctx_->memory));
-        result.emplace("day", TypedValue(date.day, ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::LocalTime: {
-        const auto &lt = expression_result.ValueLocalTime();
-        result.emplace("hour", TypedValue(lt.hour, ctx_->memory));
-        result.emplace("minute", TypedValue(lt.minute, ctx_->memory));
-        result.emplace("second", TypedValue(lt.second, ctx_->memory));
-        result.emplace("millisecond", TypedValue(lt.millisecond, ctx_->memory));
-        result.emplace("microsecond", TypedValue(lt.microsecond, ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::LocalDateTime: {
-        const auto &ldt = expression_result.ValueLocalDateTime();
-        const auto &date = ldt.date();
-        const auto &lt = ldt.local_time();
-        result.emplace("year", TypedValue(date.year, ctx_->memory));
-        result.emplace("month", TypedValue(date.month, ctx_->memory));
-        result.emplace("day", TypedValue(date.day, ctx_->memory));
-        result.emplace("hour", TypedValue(lt.hour, ctx_->memory));
-        result.emplace("minute", TypedValue(lt.minute, ctx_->memory));
-        result.emplace("second", TypedValue(lt.second, ctx_->memory));
-        result.emplace("millisecond", TypedValue(lt.millisecond, ctx_->memory));
-        result.emplace("microsecond", TypedValue(lt.microsecond, ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::ZonedDateTime: {
-        throw QueryRuntimeException("Can't coerce `{}` to Map.", expression_result.ValueZonedDateTime().ToString());
-      }
-      case TypedValue::Type::Point2d: {
-        auto const &point_2d = expression_result.ValuePoint2d();
-        result.emplace("x", TypedValue(point_2d.x(), ctx_->memory));
-        result.emplace("y", TypedValue(point_2d.y(), ctx_->memory));
-        result.emplace("srid", TypedValue(storage::CrsToSrid(point_2d.crs()).value_of(), ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Point3d: {
-        auto const &point_3d = expression_result.ValuePoint3d();
-        result.emplace("x", TypedValue(point_3d.x(), ctx_->memory));
-        result.emplace("y", TypedValue(point_3d.y(), ctx_->memory));
-        result.emplace("z", TypedValue(point_3d.z(), ctx_->memory));
-        result.emplace("srid", TypedValue(storage::CrsToSrid(point_3d.crs()).value_of(), ctx_->memory));
-        return TypedValue(result, ctx_->memory);
-      }
-      case TypedValue::Type::Graph: {
-        const auto &graph = expression_result.ValueGraph();
-
-        utils::pmr::vector<TypedValue> vertices(ctx_->memory);
-        vertices.reserve(graph.vertices().size());
-        for (const auto &v : graph.vertices()) {
-          vertices.emplace_back(TypedValue(v, ctx_->memory));
-        }
-        result.emplace("nodes", TypedValue(std::move(vertices), ctx_->memory));
-
-        utils::pmr::vector<TypedValue> edges(ctx_->memory);
-        edges.reserve(graph.edges().size());
-        for (const auto &e : graph.edges()) {
-          edges.emplace_back(TypedValue(e, ctx_->memory));
-        }
-        result.emplace("edges", TypedValue(std::move(edges), ctx_->memory));
-
-        return TypedValue(result, ctx_->memory);
-      }
-      default:
-        throw QueryRuntimeException(
-            "Only nodes, edges, maps, temporal types, points, and graphs have properties to be looked up.");
-    }
-  }
+  TypedValue Visit(AllPropertiesLookup &all_properties_lookup) override;
 
   TypedValue Visit(LabelsTest &labels_test) override {
     auto expression_result = labels_test.expression_->Accept(*this);
