+++ src/query/plan/operator.cpp
@@ -37,6 +37,7 @@
 #include "license/license.hpp"
 #include "query/auth_checker.hpp"
 #include "query/context.hpp"
+#include "query/db_accessor.hpp"
 #include "query/exceptions.hpp"
 #include "query/frontend/ast/ast.hpp"
 #include "query/frontend/semantic/symbol_table.hpp"
@@ -361,6 +362,13 @@ std::vector<Symbol> CreateExpand::ModifiedSymbols(const SymbolTable &table) cons
   symbols.emplace_back(node_info_.symbol);
   symbols.emplace_back(edge_info_.symbol);
   return symbols;
+}
+
+std::string CreateExpand::ToString() const {
+  return fmt::format("CreateExpand ({}){}[{}:{}]{}({})", input_symbol_.name(),
+                     edge_info_.direction == query::EdgeAtom::Direction::IN ? "<-" : "-", edge_info_.symbol.name(),
+                     dba_->EdgeTypeToName(edge_info_.edge_type),
+                     edge_info_.direction == query::EdgeAtom::Direction::OUT ? "->" : "-", node_info_.symbol.name());
 }
 
 CreateExpand::CreateExpandCursor::CreateExpandCursor(const CreateExpand &self, utils::MemoryResource *mem)
@@ -1107,6 +1115,15 @@ std::vector<Symbol> Expand::ModifiedSymbols(const SymbolTable &table) const {
   symbols.emplace_back(common_.node_symbol);
   symbols.emplace_back(common_.edge_symbol);
   return symbols;
+}
+
+std::string Expand::ToString() const {
+  return fmt::format(
+      "Expand ({}){}[{}{}]{}({})", input_symbol_.name(),
+      common_.direction == query::EdgeAtom::Direction::IN ? "<-" : "-", common_.edge_symbol.name(),
+      utils::IterableToString(common_.edge_types, "|",
+                              [this](const auto &edge_type) { return ":" + dba_->EdgeTypeToName(edge_type); }),
+      common_.direction == query::EdgeAtom::Direction::OUT ? "->" : "-", common_.node_symbol.name());
 }
 
 Expand::ExpandCursor::ExpandCursor(const Expand &self, utils::MemoryResource *mem)
@@ -2827,6 +2844,15 @@ UniqueCursorPtr ExpandVariable::MakeCursor(utils::MemoryResource *mem) const {
     case EdgeAtom::Type::SINGLE:
       LOG_FATAL("ExpandVariable should not be planned for a single expansion!");
   }
+}
+
+std::string ExpandVariable::ToString() const {
+  return fmt::format(
+      "{} ({}){}[{}{}]{}({})", OperatorName(), input_symbol_.name(),
+      common_.direction == query::EdgeAtom::Direction::IN ? "<-" : "-", common_.edge_symbol.name(),
+      utils::IterableToString(common_.edge_types, "|",
+                              [this](const auto &edge_type) { return ":" + dba_->EdgeTypeToName(edge_type); }),
+      common_.direction == query::EdgeAtom::Direction::OUT ? "->" : "-", common_.node_symbol.name());
 }
 
 class ConstructNamedPathCursor : public Cursor {
