+++ src/query/procedure/mg_procedure_impl.hpp
@@ -25,6 +25,7 @@
 #include "query/context.hpp"
 #include "query/frontend/ast/ast.hpp"
 
+#include "query/db_accessor.hpp"
 #include "query/procedure/cypher_type_ptr.hpp"
 #include "query/typed_value.hpp"
 #include "storage/v2/view.hpp"
