+++ tests/unit/typed_value.cpp
@@ -21,6 +21,7 @@
 #include "gtest/gtest.h"
 
 #include "disk_test_utils.hpp"
+#include "query/db_accessor.hpp"
 #include "query/graph.hpp"
 #include "query/typed_value.hpp"
 #include "storage/v2/disk/storage.hpp"
